% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splp.R
\name{boot_splp}
\alias{boot_splp}
\title{Wild cluster bootstrap function.}
\usage{
boot_splp(obj, boots)
}
\arguments{
\item{obj}{An splp object.}

\item{boots}{The number of bootstraps.}
}
\value{
Returns an H x boots data.frame, where H is the maximum horizon length for the IRF. If the original splp object is named tmp and the polynomial order is three, then the returned bootstrap df will be located under tmp$splp$order_3$info$irboot.
}
\description{
Wild cluster bootstrap function.
}
