% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splp.R
\name{splp}
\alias{splp}
\title{Local Projections Estimator for Smooth Local Projections.}
\usage{
splp(
  fml,
  paneldf,
  type,
  q = 3,
  r = 2,
  lambda = 10^seq(-3, 3, length.out = 100),
  boots = 0,
  clustervar = NULL,
  targetvar = NULL,
  ci = 0.95,
  m = 1000
)
}
\arguments{
\item{fml}{Formula object with the same notation as in the package fixest. For example, to run a panel local projection ten units out with two-way fixed effects for up to ten horizons out, one would write fml = f(y,0:10) ~ x | id + time. If there is an instrumental variable write fml = f(y,0:10) ~ 1 | id + time | x ~ z. Please consult the fixest package for more details.}

\item{paneldf}{A panel dataframe created with the fixest function fixest::panel(). You must specify the original dataframe in the first argument and the panel identifiers in the second. Consult the fixest package for more details, especially if your time identifier is not integer-spaced.}

\item{type}{Takes value "splp" for SPLP or "plp" for PLP.}

\item{q}{The number of knots in the B-spline basis function. Defaults to q = 3.}

\item{r}{Specifies the polynomial order r + 1 that splp tries to discipline the IRF to. For example, a linear IRF would take r = 2, which is the default value. The function can take a vector of polynomials. For example, set r = c(2,3,4) to get a linear IRF, a quadratic IRF, and a cubic IRF.}

\item{lambda}{Vector of penalty parameters lambda to try. Defaults to a logarithmic grid from 10^{-3} to 10^3 with fifty gridpoints.}

\item{boots}{Number of bootstraps to run. Defaults to zero.}

\item{clustervar}{Which variable to cluster around when doing a wild cluster bootstrap. Defaults to the individual identifier.}

\item{targetvar}{The right-hand side variable for which we want to smooth the impulse response function.}

\item{ci}{Confidence interval with default value 0.95 (for a 95\% confidence interval).}

\item{m}{Choice for the number of random vectors in stochastic trace approximation. Default value is 1000.}
}
\value{
Returns an splp object as a list with three entries. The first, named plp, contains IRFs for the standard panel local projections estimator. The second, named info, contains basic info about the object (such as identifying variables). The third entry contains a list of splp objects. Each is named according to its polynomial order. If you set r = c(2,3,4), then there will be three objects named order_1, order_2, and order_3. Each underlying object has the relevant IRFs as well as a plot comparing its IRF to the standard plp IRF.
}
\description{
Constructs IRFs for both unsmoothed and smoothed local projections disciplined to a polynomial of your choice. The smoothing estimator is based on Barnichon & Brownlees (2019, ReStat) and extended to panel data. It can handle both OLS and IV methods for panel local projections as well as a variety of fixed effects.
}
