% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splp.R
\name{splp_ci_out}
\alias{splp_ci_out}
\title{Plots IRFs and computes confidence interval from bootstraps.}
\usage{
splp_ci_out(obj, ci, boots)
}
\arguments{
\item{obj}{An splp object.}

\item{ci}{A confidence interval.}

\item{boots}{The number of bootstraps.}
}
\value{
Returns a smoothed IRF plot and a plot comparing the smoothed to the unsmoothed plots.  If the original splp object is named tmp and the polynomial order is three, then the returned combined plot will be located under tmp$splp$order_3[\link{2}] and the smoothed IRF alone will be under tmp$splp$order_3[\link{2}]$smoothplot. If boots = 0, then only point estimates will be returned.
}
\description{
Plots IRFs and computes confidence interval from bootstraps.
}
